+++ src-docs/public/dialog-template.html
@@ -7,7 +7,7 @@
 </mdw-content>
 
 <mdw-footer>
-  <mdw-button class="mdw-error" onclick="MDWDialog.close()">cancel</mdw-button>
+  <mdw-button onclick="MDWDialog.close()">cancel</mdw-button>
   <!--- MDWDialog.close(true) will reolve the dialgo promise with true -->
   <mdw-button onclick="MDWDialog.close(true)">ok</mdw-button>
 </mdw-footer>
