+++ src-docs/public/surfaces/one.html
@@ -1,0 +1,22 @@
+<mdw-header>
+  <section class="mdw-flex">
+    <div class="mdw-title">Main title</div>
+    <div class="mdw-subtitle">Sub title</div>
+  </section>
+
+  <section>
+    <mdw-button onclick="MDWSurface.close(); event.stopPropagation()" class="mdw-icon">
+      <mdw-icon>close</mdw-icon>
+    </mdw-button>
+  </section>
+</mdw-header>
+
+<mdw-content>
+  <mdw-list>
+    <mdw-list-item>One</mdw-list-item>
+    <mdw-list-item>Two</mdw-list-item>
+    <mdw-list-item>Three</mdw-list-item>
+    <mdw-list-item>Four</mdw-list-item>
+    <mdw-list-item>Five</mdw-list-item>
+  </mdw-list>
+</mdw-content>
